+++ source/common/router/header_parser.cc
@@ -37,19 +37,16 @@ std::string unescape(absl::string_view sv) { return absl::StrReplaceAll(sv, {{"%
 // RequestInfoHeaderFormatter.
 HeaderFormatterPtr
 parseInternal(const envoy::api::v2::core::HeaderValueOption& header_value_option) {
-  // We reject :path rewriting, there is already a well defined mechanism to
+  const std::string& key = header_value_option.header().key();
+  // PGV constraints provide this guarantee.
+  ASSERT(!key.empty());
+  // We reject :path/:authority rewriting, there is already a well defined mechanism to
   // perform this in the RouteAction, and doing this via request_headers_to_add
   // will cause us to have to worry about interaction with other aspects of the
-  // RouteAction, e.g. prefix rewriting.
-  const std::string& key = header_value_option.header().key();
-  if (key == Http::Headers::get().Path.get()) {
-    throw EnvoyException(
-        ":path rewriting may not be done via header addition, use prefix_rewrite instead.");
-  }
-  // Same goes for host in the interest of economy of mechanism.
-  if (key == Http::Headers::get().Host.get()) {
-    throw EnvoyException(
-        ":authority rewriting may not be done via header addition, use host_rewrite instead.");
+  // RouteAction, e.g. prefix rewriting. We also reject other :-prefixed
+  // headers, since it seems dangerous and there doesn't appear a use case.
+  if (key[0] == ':') {
+    throw EnvoyException(":-prefixed headers may not be modified");
   }
 
   const bool append = PROTOBUF_GET_WRAPPED_OR_DEFAULT(header_value_option, append, true);